fix: sets the security env variables as first
